+++ api/src/main/java/com/opensourcereader/api/facade/auth/AuthFacadeService.java
@@ -1,5 +1,6 @@
 package com.opensourcereader.api.facade.auth;
 
+import com.opensourcereader.core.user.service.UserSignUpService;
 import java.time.Instant;
 
 import org.springframework.stereotype.Service;
@@ -8,7 +9,6 @@ import com.opensourcereader.api.controller.auth.request.SignUpRequest;
 import com.opensourcereader.core.security.dto.UserConnection;
 import com.opensourcereader.core.user.dto.UserSignUpCommand;
 import com.opensourcereader.core.user.entity.User;
-import com.opensourcereader.core.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -16,12 +16,12 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class AuthFacadeService {
 
-  private final UserService userService;
+  private final UserSignUpService signUpService;
 
   public UserConnection signup(SignUpRequest request) {
     UserSignUpCommand command =
         new UserSignUpCommand(request.email(), request.password(), request.loginName(), null, null);
-    User user = userService.signup(command);
+    User user = signUpService.signup(command);
     return new UserConnection(user, Instant.now());
   }
 }
