+++ api/src/main/java/com/opensourcereader/api/facade/user/UserFacadeService.java
@@ -1,12 +1,12 @@
 package com.opensourcereader.api.facade.user;
 
+import com.opensourcereader.core.user.service.UserRetrieveService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.opensourcereader.api.controller.user.request.UserGetRequest;
 import com.opensourcereader.core.user.dto.UserDto;
 import com.opensourcereader.core.user.entity.User;
-import com.opensourcereader.core.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -14,11 +14,11 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class UserFacadeService {
 
-  private final UserService userService;
+  private final UserRetrieveService userRetrieveService;
 
   @Transactional(readOnly = true)
   public UserDto getUser(final UserGetRequest req) {
-    User user = userService.findByNickname(req.nickname());
+    User user = userRetrieveService.findByNickname(req.nickname());
 
     return UserDto.from(user);
   }
