+++ api/src/test/java/com/opensourcereader/api/facade/github/GithubFacadeServiceTest.java
@@ -22,7 +22,7 @@ import com.opensourcereader.api.client.request.GithubRepoRequest;
 import com.opensourcereader.core.analysis.entity.OpenSourceRepo;
 import com.opensourcereader.core.analysis.service.impl.LocalOpenSourceRepoService;
 import com.opensourcereader.core.board.service.IssueSyncService;
-import com.opensourcereader.core.user.service.UserServiceImpl;
+import com.opensourcereader.core.user.service.UserRetrieveService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -45,7 +45,7 @@ public class GithubFacadeServiceTest {
 
   @Autowired private LocalOpenSourceRepoService openSourceRepoService;
 
-  @Autowired private UserServiceImpl userService;
+  @Autowired private UserRetrieveService userRetrieveService;
 
   @Autowired private IssueSyncService issueSyncService;
