+++ core/src/main/java/com/opensourcereader/core/user/service/UserSignUpService.java
@@ -1,29 +1,24 @@
 package com.opensourcereader.core.user.service;
 
-import java.util.UUID;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.opensourcereader.core.security.dto.UserInfo;
 import com.opensourcereader.core.user.dto.GithubUserCommand;
 import com.opensourcereader.core.user.dto.UserSignUpCommand;
 import com.opensourcereader.core.user.entity.Role;
 import com.opensourcereader.core.user.entity.User;
-import com.opensourcereader.core.user.exception.UserNotFoundException;
 import com.opensourcereader.core.user.repository.UserRepository;
-
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
+public class UserSignUpService {
 
   private final UserRepository userRepository;
   private final PasswordEncoder passwordEncoder;
 
-  @Override
   @Transactional
   public User signup(UserSignUpCommand command) {
     User user =
@@ -39,12 +34,11 @@ public class UserServiceImpl implements UserService {
     return userRepository.save(user);
   }
 
-  @Override
   @Transactional
   public User signup(UserInfo userInfo) {
     User user =
         userRepository
-            .findFirstByLoginName(userInfo.loginName())
+            .findByProviderId(userInfo.providerId())
             .orElseGet(() -> User.from(userInfo));
 
     user.updateAvatar(userInfo.avatarUrl());
@@ -57,12 +51,11 @@ public class UserServiceImpl implements UserService {
     return userRepository.save(user);
   }
 
-  @Override
   @Transactional
   public User guest(GithubUserCommand command) {
     User user =
         userRepository
-            .findFirstByLoginName(command.loginName())
+            .findByProviderId(command.providerId())
             .orElseGet(() -> User.newGuest(command));
     user.updateAvatar(command.avatarUrl());
     user.linkSocialProvider(command.providerId());
@@ -70,23 +63,5 @@ public class UserServiceImpl implements UserService {
       user.updatePassword(UUID.randomUUID().toString());
     }
     return userRepository.save(user);
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  public User findByProviderId(Long providerId) {
-    return userRepository.findByProviderId(providerId).orElseThrow(UserNotFoundException::new);
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  public User findByNickname(String loginName) {
-    return userRepository.findFirstByLoginName(loginName).orElseThrow(UserNotFoundException::new);
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  public User findByEmail(String email) {
-    return userRepository.findFirstByEmail(email).orElseThrow(UserNotFoundException::new);
   }
 }
